Python/benchmark/pyproject.toml
[project]
name = "benchmark"
version = "0.2.0"
description = "A set of micro-benchmarks for the Python programming language"
authors = [
    { name = "Kay-Uwe Kirstein", email = "kay-uwe@kirsteinhome.ch" }
]
license = "MIT"
requires-python = ">=3.14"

dependencies = [
    # Add runtime dependencies here
    "pillow>=12.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[project.scripts]
benchmark = "benchmark.runner:run"

[build-system]
requires = ["uv-build>=0.9.27,<0.10.0"]
build-backend = "uv_build"

